sending full struct

+++ rt/include/rt_spi.h
@@ -162,7 +162,8 @@ typedef struct {
   float tau_abad_ff[4];
   float tau_hip_ff[4];
   float tau_knee_ff[4];
-  int32_t flags[4];
+  // int32_t flags[4];
+  float flags[4];
   int32_t checksum;
 
 } spine_biqu_cmd_t;
